Use callbacks instead of events
For icon updating, we allow the provider to explicitly actively update
the icon using the callback instead of using the event listeners
implemented in the Tab class. By doing this, the provider has the
freedom to update the icon any time it wants.

For tab reloading, we move the logic to the Tab class so that the
provider doesn't have to use the Chrome API directly and allow the
provider to reload the tab using the callback in the same way as in icon
updating.