Add tests for Cloudflare provider
The issuance involves handleBeforeRequest listener.
1. Firstly, the listener check if the request looks like the one that we
should send an issuance request.
2. If it passes the check, the listener returns the cancel command to
cancel the request. If not, it returns nothing and let the request
continue.
3. At the same time the listener returns, it calls a private method
"issue" to send an issuance request to the server and the method return
an array of issued tokens.
4. The listener stored the issued tokens in the storage.
5. The listener reloads the tab to get the proper web page for the tab.

The redemption involves handleHeadersReceived and handleBeforeSendHeaders
listeners. In handleHeadersReceived listener,
1. Firstly, the listener check if the response is the challenge page and
it supports Privacy Pass redemption.
2. If it passes the check, the listener gets a token from the storage to
redeem.
3. The listener sets "redeemInfo" property which includes the request id
and the mentioned token. The property will be used by
handleBeforeSendHeaders to redeem the token.
4. The listener returns the redirect command so that the browser will
send the same request again with the token attached.

In handleBeforeSendHeaders,
1. The listener will check if the provided request id matches the
request id in "redeemInfo". If so, it means that the request is from the
redirect command returned by handleHeadersReceived. If not, it returns
nothing and let the request continue.
2. If it passes the check, the listener attaches the token from
"redeemInfo" in the "challenge-bypass-token" HTTP header and clears the
"redeemInfo" property because "redeemInfo" is used already.