detect and fix incorrect requests on Cloudflare issuing domain
This methodology is aggressive.

It doesn't aim to fix the race-condition
that exists on the Cloudflare website;
only Cloudflare can do that directly.

It provides a workaround to deal with both possible outcomes.
One outcome is desirable, and everything works as it should.
The other outcome is unexpected, and this commit:
 * detects the incorrect request
 * cancels the incorrect request
 * cherry picks data from its headers
 * constructs the correct request to issue tokens

This methodology is very similar to how a correct request is processed.
The difference is where the data originates,
which is then used to construct the request to issue tokens.

a "correct" request:
====================
  curl 'https://captcha.website/?__cf_chl_captcha_tk__=xxx' \
    -H 'Content-Type: application/x-www-form-urlencoded' \
    -H 'Referer: https://captcha.website/' \
    --data-raw 'md=xxx&r=xxx&cf_captcha_kind=h&vc=xxx&captcha_vc=xxx&captcha_answer=xxx&cf_ch_verify=plat&h-captcha-response=captchka'

an "incorrect" request:
=======================
  curl 'https://captcha.website/' \
    -H 'content-type: application/x-www-form-urlencoded' \
    -H 'referer: https://captcha.website/?__cf_chl_tk=xxx' \
    --data-raw 'md=xxx&r=xxx&cf_captcha_kind=h&vc=xxx&captcha_vc=xxx&captcha_answer=xxx&cf_ch_verify=plat&h-captcha-response=captchka'

In a "correct" request:
 * querystring of url contains value for key: '__cf_chl_captcha_tk__'

In an "incorrect" request,
 * querystring of 'referer' header contains value for key: '__cf_chl_tk'
 * the "correct" url can be constructed by adding this value to the querystring key: '__cf_chl_captcha_tk__'
 * apparently, they are interchangeable!
   - the resulting request to issue tokens works 100% of the time