Change the issue request flow for Cloudflare
Previously we used axios to send an issuance request in
handleBeforeRequest, but now we will send the request in
handleBeforeSendHeaders because we need to read the Referer header
first.

After reading the Referer header, we extract the __cf_chl_tk query param
in the Referer url and add the __cf_chl_f_tk query param with that token
in the issuance request URL.

We need a new "issueInfo" property to remember the request body we can
read in handleBeforeRequest and then it will be read in
handleBeforeSendHeaders so that in handleBeforeSendHeaders we will have
both the body and the Referer header.