Add notice for end of support
Privacy Pass has changed since the original PoPETS version. Moreover,
CAPTCHA providers have moved from that version to follow IETF version
more closely. They don't operate with a stored token model, but with an
interactive token issuance and redemption, with a split attester/issuer
model.
This commit updates the README to highlight the end of active token
issuer/redeemer for this specific version of the protocol.